requirements.txt
fastapi
redis
uvicorn
tensorflow
numpy
pillow
python-multipart
python-dotenv
matplotlib
scikit-learn
opencv-python
celery
google-cloud-aiplatform
protobuf